Simplify key bind setting
